Update User.cs
making the Id bigger, since int32 is too small

+++ oAuthTwitterWrapper/JsonTypes/User.cs
@@ -14,7 +14,7 @@ namespace OAuthTwitterWrapper.JsonTypes
     {
 
 		[JsonProperty("id")]
-		public int Id { get; set; }
+		public double Id { get; set; }
 
 		[JsonProperty("id_str")]
 		public string IdStr { get; set; }
